refactor: model 타입 구조 개선
