+++ src/api/modelApi.ts
@@ -1,6 +1,6 @@
-import api from "./axios";
+import { api } from "./axios";
+import type { ModelListResponse, ModelDetailResponse } from "@/types/model";
 
-// 공통 API 응답 형태
 export interface ApiError {
   code: string;
   message: string;
@@ -13,24 +13,21 @@ export interface ApiResponse<T> {
   error?: ApiError | null;
 }
 
-// 모델 객체
-export interface ModelObject {
-  id: number;
-  name: string;
-  description: string;
-  thumbnailUrl: string;
-  mainTheory: string;
-  usage: string;
-}
-
 // 모델 객체 목록
-export const getModelObjects = async (): Promise<
-  ApiResponse<ModelObject[]>
-> => {
-  const { data } = await api.get<ApiResponse<ModelObject[]>>("/model-objects");
+export const getModelObjects = async (): Promise<ModelListResponse> => {
+  const { data } = await api.get<ModelListResponse>("/model-objects");
+  return data;
+};
+
+// 모델 객체 상세 조회
+export const getDetailModel = async (
+  id: number,
+): Promise<ModelDetailResponse> => {
+  const { data } = await api.get<ModelDetailResponse>(`/model-objects/${id}`);
   return data;
 };
 
 export default {
   getModelObjects,
+  getDetailModel,
 };
