+++ src/types/model.ts
@@ -1,3 +1,6 @@
+// src/types/model.ts
+import type { ApiResponse } from "@/api/modelApi";
+
 /**
  * model.ts
  *
@@ -7,6 +10,7 @@
  * - MODEL_NAMES: 허용되는 모델 이름 목록
  * - ModelType: MODEL_NAMES로부터 나온 모델 타입
  */
+
 export const MODEL_NAMES = [
   "Robot-Gripper",
   "Suspension",
@@ -15,3 +19,18 @@ export const MODEL_NAMES = [
 ] as const;
 
 export type ModelType = (typeof MODEL_NAMES)[number];
+
+// 모델 객체
+export interface ModelObject {
+  id: number;
+  name: ModelType;
+  description: string;
+  thumbnailUrl: string;
+  mainTheory: string;
+  usage: string;
+  viewData?: string;
+}
+
+// API 응답을 위한 타입 확장
+export type ModelListResponse = ApiResponse<ModelObject[]>;
+export type ModelDetailResponse = ApiResponse<ModelObject>;
